+++ src/hr/avrbanac/openglplayground/renderers/MasterRenderer.java
@@ -1,6 +1,7 @@
 
-package hr.avrbanac.openglplayground.renderengine;
+package hr.avrbanac.openglplayground.renderers;
 
+import static hr.avrbanac.openglplayground.Globals.*;
 import hr.avrbanac.openglplayground.entities.Camera;
 import hr.avrbanac.openglplayground.entities.Entity;
 import hr.avrbanac.openglplayground.entities.Light;
@@ -37,9 +38,7 @@ public class MasterRenderer {
     private List<Terrain> terrains = new ArrayList<>();
 
     public MasterRenderer(int width, int height, float fov) {
-        // don't render faces that will never be seen
-        GL11.glEnable(GL11.GL_CULL_FACE);
-        GL11.glCullFace(GL11.GL_BACK);
+        enableCulling();
         
         projectionMatrix = Matrix4f.projection(width, height, fov);
         
@@ -50,16 +49,29 @@ public class MasterRenderer {
         terrainRenderer = new TerrainRenderer(terrainShader, projectionMatrix);
     }
     
+    public static void enableCulling() {
+        // don't render faces that will never be seen
+        GL11.glEnable(GL11.GL_CULL_FACE);
+        GL11.glCullFace(GL11.GL_BACK);
+    }
+    
+    public static void disableCulling() {
+        // don't render faces that will never be seen
+        GL11.glDisable(GL11.GL_CULL_FACE);
+    }
+    
     public void render(Light sun, Camera camera) {
         prepare();
         
         shader.start();
+        shader.loadSkyColor(SKY_RED, SKY_GREEN, SKY_BLUE);
         shader.loadLight(sun);
         shader.loadViewMatrix(camera);
         renderer.render(entities);
         shader.stop();
         
         terrainShader.start();
+        terrainShader.loadSkyColor(SKY_RED, SKY_GREEN, SKY_BLUE);
         terrainShader.loadLight(sun);
         terrainShader.loadViewMatrix(camera);
         terrainRenderer.render(terrains);
@@ -92,7 +104,7 @@ public class MasterRenderer {
         // clear the color from previous frame
         GL11.glClear(GL11.GL_COLOR_BUFFER_BIT | GL11.GL_DEPTH_BUFFER_BIT);
         // clear everything with background color (black in this case)
-        GL11.glClearColor(0, 0, 0, 1);
+        GL11.glClearColor(SKY_RED, SKY_GREEN, SKY_BLUE, 1);
     }
     
     public void cleanUp() {