+++ src/hr/avrbanac/openglplayground/textures/ModelTexture.java
@@ -9,8 +9,15 @@ package hr.avrbanac.openglplayground.textures;
  */
 public class ModelTexture {
     private int textureID;
+    
     private float shineDamper = 1;
     private float reflectivity = 0;
+    
+    // this is used for turning off face culling 
+    private boolean transparency = false;
+    
+    // this is used for lighting simple terrain objects which use very small number of quad planes (normals)
+    private boolean useFakeLighting = false;
     
     public ModelTexture(int textureID) {
         this.textureID = textureID;
@@ -35,5 +42,22 @@ public class ModelTexture {
     public void setReflectivity(float reflectivity) {
         this.reflectivity = reflectivity;
     }
+
+    public boolean isTransparency() {
+        return transparency;
+    }
+
+    public void setTransparency(boolean transparency) {
+        this.transparency = transparency;
+    }
+
+    public boolean isUseFakeLighting() {
+        return useFakeLighting;
+    }
+
+    public void setUseFakeLighting(boolean useFakeLighting) {
+        this.useFakeLighting = useFakeLighting;
+    }
+    
     
 }